controller/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
openai==1.12.0
python-multipart==0.0.6
PyYAML==6.0.1
httpx==0.25.2
